refactor: improve logging

+++ pkg/cluster/network/delete.go
@@ -7,10 +7,16 @@ import (
 
 func Delete(ctx clustercontext.ClusterContext) {
 	network := Get(ctx)
+	if network == nil {
+		return
+	}
+
+	log.Println("Deleting network - ", network.Name)
 
 	_, err := ctx.Client.Network.Delete(ctx.Context, network)
 	if err != nil {
-		log.Fatalf("error deleting network: %s", err)
+		log.Println("error deleting network: ", err)
 	}
-	ctx.Network = nil
+
+	log.Println("Network deleted - ", network.Name)
 }
